+++ spec/jsonapi/resources/matchers/have_model_name_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe JSONAPI::Resources::Matchers::HaveModelName do
 
   describe "#failure_message" do
     let(:author) { Author.new }
-    let(:resource) { AuthorResource.new(author) }
+    let(:resource) { AuthorResource.new(author, {}) }
     let(:matcher) { described_class.new("Friend") }
     subject(:failure_message) { matcher.failure_message }