+++ spec/jsonapi/resources/matchers_spec.rb
@@ -9,21 +9,21 @@ describe JSONAPI::Resources::Matchers do
   describe "resource matchers", type: :resource do
     describe "attributes" do
       let(:author) { Author.new }
-      subject(:resource) { AuthorResource.new(author) }
+      subject(:resource) { AuthorResource.new(author, {}) }
       it { is_expected.to have_attribute(:name) }
       it { is_expected.to_not have_attribute(:created_at) }
     end
 
     describe "filters" do
       let(:author) { Author.new }
-      subject(:resource) { AuthorResource.new(author) }
+      subject(:resource) { AuthorResource.new(author, {}) }
       it { is_expected.to filter(:name) }
       it { is_expected.to_not filter(:created_at) }
     end
 
     describe "have many" do
       let(:author) { Author.new }
-      subject(:resource) { AuthorResource.new(author) }
+      subject(:resource) { AuthorResource.new(author, {}) }
       it { is_expected.to have_many(:books) }
       it { is_expected.to have_many(:libros).with_class_name("Book") }
       it { is_expected.to have_many(:libros).with_relation_name(:books) }
@@ -32,7 +32,7 @@ describe JSONAPI::Resources::Matchers do
 
     describe "have one" do
       let(:book) { Book.new }
-      subject(:resource) { LibroResource.new(book) }
+      subject(:resource) { LibroResource.new(book, {}) }
       it { is_expected.to have_one(:author) }
       it { is_expected.to have_one(:author).with_class_name("Writer") }
       it { is_expected.to have_one(:author).with_relation_name(:writer) }
@@ -40,14 +40,14 @@ describe JSONAPI::Resources::Matchers do
 
     describe "model name" do
       let(:book) { Book.new }
-      subject(:resource) { LibroResource.new(book) }
+      subject(:resource) { LibroResource.new(book, {}) }
       it { is_expected.to have_model_name("Book") }
       it { is_expected.to_not have_model_name("Libro") }
     end
 
     describe "primary key" do
       let(:author) { Author.new }
-      subject(:resource) { AuthorResource.new(author) }
+      subject(:resource) { AuthorResource.new(author, {}) }
       it { is_expected.to have_primary_key(:name) }
       it { is_expected.to_not have_primary_key(:id) }
     end